requirements.txt
openai-whisper
pydub
requests
google-cloud-speech